Enhances call notification and handling
Improves the incoming call notification process by launching the CallingActivity upon user interaction with the notification, instead of automatically.

This change also refactors the notification to use NotificationCompat for better compatibility and includes explicit intents for answering and declining calls. It handles edge cases, such as when the caller name or number is unavailable.

Additionally, it requests overlay permissions on Android devices running API level 23 (Marshmallow) or higher, which is necessary to display the calling activity on top of other applications.